+++ src/main/java/org/opentripplanner/graph_builder/module/osm/OSMDatabase.java
@@ -9,6 +9,7 @@ import gnu.trove.map.TLongObjectMap;
 import gnu.trove.map.hash.TLongObjectHashMap;
 import gnu.trove.set.TLongSet;
 import gnu.trove.set.hash.TLongHashSet;
+import java.time.ZoneId;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -18,6 +19,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
+import java.util.function.Supplier;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
@@ -128,10 +130,20 @@ public class OSMDatabase {
    */
   public boolean noZeroLevels = true;
   private final Set<String> boardingAreaRefTags;
+  private final Supplier<ZoneId> timeZone;
 
-  public OSMDatabase(DataImportIssueStore issueStore, Set<String> boardingAreaRefTags) {
+  public OSMDatabase(
+    DataImportIssueStore issueStore,
+    Set<String> boardingAreaRefTags,
+    Supplier<ZoneId> timeZoneId
+  ) {
     this.issueStore = issueStore;
     this.boardingAreaRefTags = boardingAreaRefTags;
+    this.timeZone = timeZoneId;
+  }
+
+  public OSMDatabase(DataImportIssueStore issueStore, Set<String> boardingAreaRefTags) {
+    this(issueStore, boardingAreaRefTags, () -> ZoneId.of("UTC"));
   }
 
   public OSMNode getNode(Long nodeId) {
@@ -1027,7 +1039,8 @@ public class OSMDatabase {
             relation.getTag("day_on"),
             relation.getTag("day_off"),
             relation.getTag("hour_on"),
-            relation.getTag("hour_off")
+            relation.getTag("hour_off"),
+            timeZone
           );
       } catch (NumberFormatException e) {
         LOG.info("Unparseable turn restriction: " + relation.getId());