+++ src/main/java/org/opentripplanner/graph_builder/module/osm/OpenStreetMapModule.java
@@ -17,6 +17,7 @@ import java.util.Map;
 import java.util.OptionalInt;
 import java.util.Set;
 import java.util.stream.Collectors;
+import javax.annotation.Nullable;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
@@ -141,11 +142,13 @@ public class OpenStreetMapModule implements GraphBuilderModule {
   private final Graph graph;
   private final ZoneId timeZoneId;
 
+  private boolean hasWarnedAboutMissingTimeZone = false;
+
   public OpenStreetMapModule(
     Collection<OpenStreetMapProvider> providers,
     Set<String> boardingAreaRefTags,
     Graph graph,
-    ZoneId timeZoneId,
+    @Nullable ZoneId timeZoneId,
     DataImportIssueStore issueStore
   ) {
     this.providers = List.copyOf(providers);
@@ -175,15 +178,6 @@ public class OpenStreetMapModule implements GraphBuilderModule {
     this.maxAreaNodes = config.maxAreaNodes;
   }
 
-  public OpenStreetMapModule(
-    OpenStreetMapProvider provider,
-    Graph graph,
-    ZoneId timeZoneId,
-    DataImportIssueStore issueStore
-  ) {
-    this(List.of(provider), Set.of(), graph, timeZoneId, issueStore);
-  }
-
   /**
    * Set the way properties from a {@link WayPropertySetSource} source.
    *
@@ -198,8 +192,13 @@ public class OpenStreetMapModule implements GraphBuilderModule {
   @Override
   public void buildGraph() {
     this.osmOpeningHoursParser =
-      new OSMOpeningHoursParser(graph.getOpeningHoursCalendarService(), timeZoneId, issueStore);
-    OSMDatabase osmdb = new OSMDatabase(issueStore, boardingAreaRefTags);
+      new OSMOpeningHoursParser(
+        graph.getOpeningHoursCalendarService(),
+        this::getTimeZone,
+        issueStore
+      );
+
+    OSMDatabase osmdb = new OSMDatabase(issueStore, boardingAreaRefTags, this::getTimeZone);
     Handler handler = new Handler(graph, osmdb);
     for (OpenStreetMapProvider provider : providers) {
       LOG.info("Gathering OSM from provider: " + provider);
@@ -1679,5 +1678,17 @@ public class OpenStreetMapModule implements GraphBuilderModule {
       }
       return vertices.get(level);
     }
+  }
+
+  private ZoneId getTimeZone() {
+    if (timeZoneId == null) {
+      if (!hasWarnedAboutMissingTimeZone) {
+        hasWarnedAboutMissingTimeZone = true;
+        LOG.warn(
+          "Missing time zone - time-restricted entities will not be created, please configure it in the build-config.json"
+        );
+      }
+    }
+    return timeZoneId;
   }
 }